Inline size of value and remove user metadata and CAS.
User metadata should not take space if not required by app. CAS
should not take space if not required by app, and should not mandate
using a tick counter to perform CAS. Removing these until a tighter
implementation is added.

Inlining the size of value removes a mutex from the skiplist node,
which significantly speeds up insertion and search time.